slack-bot-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
slack-bolt==1.20.0
google-cloud-aiplatform==1.70.0
google-cloud-bigquery==3.25.0
google-cloud-secret-manager==2.20.0
redis==5.0.8
pydantic-settings==2.5.2
httpx==0.27.2
